WIP Pull type querying with substitution into its own (file and) typeclass.
Why?  Because checking types is restricted to rename-only
substitutions (which the checker itself generates), but we want to be
able to query types through a more meaningful substitution, such as
the one carried by SimplifyM.

The ultimate purpose is to be able to (efficiently) query types in the
input space of a pass (without having to subtitute the whole
atom/expr/block first).